requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandasql>=0.7.3
sqlalchemy>=2.0.0
tabulate>=0.9.0
